Add seeker stop condition for unsolvable mazes
Implemented detection for when all walkable maze tiles have been visited that happens when the seeker finds a valid square to move to. Variable *allVisited* gets set to true and if every square is visited during the 2D vector search, the main movement logic loop doesn't run anymore, making the seeker halt to a stop. At the start *allVisited* is set to false, making it possible to run the movement logic loop at least once.