Create sudoku game / generator
Made a sudoku game that generates a random sudoku puzzle. Grids with multiple solutions can still be generated, though. Had fun working with the UI.
Controls: hover over a cell and press a number on the num row to insert that number into an empty cell. Backspace to delete the number.